requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic[email]
pymysql
python-dotenv